--- library/objective-c/Envoy-Bridging-Header.h
@@ -1,4 +1,0 @@
-#import "library/objective-c/EnvoyEngine.h"
-#import "library/objective-c/EnvoyEngineImpl.h"
-#import "library/objective-c/EnvoyHttpStream.h"
-#import "library/objective-c/EnvoyTypes.h"
