+++ library/objective-c/EnvoyEngine.m
@@ -1,9 +1,39 @@
-#import "library/objective-c/EnvoyHttpStream.h"
-
-#import "library/common/include/c_types.h"
+#import "library/objective-c/EnvoyEngine.h"
 #import "library/common/main_interface.h"
-
+#import "library/common/include/c_types.h"
 #import <stdatomic.h>
+
+@implementation EnvoyEngineError
+@end
+
+@implementation EnvoyObserver
+@end
+
+@implementation EnvoyEngineImpl
+
+#pragma mark - class methods
++ (EnvoyStatus)runWithConfig:(NSString *)config {
+  return [self runWithConfig:config logLevel:@"info"];
+}
+
++ (EnvoyStatus)runWithConfig:(NSString *)config logLevel:(NSString *)logLevel {
+  // Envoy exceptions will only be caught here when compiled for 64-bit arches.
+  // https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/Exceptions/Articles/Exceptions64Bit.html
+  @try {
+    return (EnvoyStatus)run_engine(config.UTF8String, logLevel.UTF8String);
+  } @catch (...) {
+    NSLog(@"Envoy exception caught.");
+    [NSNotificationCenter.defaultCenter postNotificationName:@"EnvoyException" object:self];
+    return EnvoyStatusFailure;
+  }
+}
+
++ (void)setupEnvoy {
+  setup_envoy();
+}
+
+@end
+
 
 #pragma mark - utility functions to move elsewhere
 typedef struct {
