--- library/objective-c/EnvoyTypes.m
@@ -1,7 +1,0 @@
-#import "library/objective-c/EnvoyTypes.h"
-
-@implementation EnvoyEngineError
-@end
-
-@implementation EnvoyObserver
-@end
